d_a_npc_aj1: decideType s32->u8 (retail caller clrlwi.24 in _create) — _create 98.85->99.26; decideType stays 98.13 (mr-vs-clrlwi bool mask, accept-class); remaining _create diffs = l_HIO cull-size pool-shift (values match) + a_siz_tbl anchor; fresh .o postdates

+++ include/d/actor/d_a_npc_aj1.h
@@ -58,7 +58,7 @@ public:
     void lookBack();
     s32 chkAttention();
     void setAttention(bool);
-    s32 decideType(int);
+    u8 decideType(int);
     void cut_init_AJ1_TLK();
     s32 cut_move_AJ1_TLK();
     void cut_init_INI_ANGRY();
